+++ Assets/Scripts/Runtime/Controllers/BattleCurtain/BattleCurtainController.cs
@@ -1,4 +1,6 @@
-namespace Akashic.Runtime.Controllers.LoadingCurtain
+using Akashic.Runtime.Common;
+
+namespace Akashic.Runtime.Controllers.BattleCurtain
 {
     internal sealed class BattleCurtainController : CurtainController
     {
