+++ Assets/Scripts/Runtime/Controllers/LoadingCurtain/LoadingCurtainController.cs
@@ -1,3 +1,5 @@
+using Akashic.Runtime.Common;
+
 namespace Akashic.Runtime.Controllers.LoadingCurtain
 {
     internal sealed class LoadingCurtainController : CurtainController
