+++ Assets/Scripts/Runtime/Controllers/PreloadCurtain/PreloadCurtainController.cs
@@ -1,4 +1,6 @@
-namespace Akashic.Runtime.Controllers.LoadingCurtain
+using Akashic.Runtime.Common;
+
+namespace Akashic.Runtime.Controllers.PreloadCurtain
 {
     internal sealed class PreloadCurtainController : CurtainController
     {
